refactored searchProjects
