--- src/pages/projContracts/api/searchProjects.test.ts
@@ -1,9 +1,0 @@
-import { searchProjects } from './searchProjects';
-
-test('searchProjects', async ()=>{
-
-  const result = await searchProjects('Lorenz');
-
-  expect(result).toMatchSnapshot();
-
-});
