+++ src/pages/projContracts/index.ts
@@ -1,4 +1,3 @@
-export * from './api';
 export * from './form';
 export * from './FormikContractPreview';
 export * from './FormContractPreview';
