+++ src/pages/projContracts/parts/SearchProjField.tsx
@@ -2,9 +2,9 @@ import { Autocomplete, TextField, Stack } from '@mui/material';
 import { useField, useFormikContext } from 'formik';
 import { useEffect, useState } from 'react';
 import { useLazyEffect } from '../../../hooks';
-import { searchProjects } from '../api';
 import { Caption } from '../../../components/ui/typographies';
 import { getFieldName, TypeOfForm } from '../form';
+import { searchProjects } from '../../../api/kintone/projects';
 
 type Opt = {
   id: string,
