+++ src/pages/projProspect/index.ts
@@ -1,4 +1,3 @@
-export * from './api';
 export * from './form';
 export * from './FormikProjProspect';
 export * from './FormProjProspect';
